Tach ket noi DB va API thanh file rieng
